Habits CRUD Operations Implemented

+++ service/habitsService.js
@@ -1,27 +1,34 @@
 const habitDAO = require('../repository/habitsDAO');
+const ElementNotFoundException = require('../exceptions/elementNotFoundException');
 
-exports.fetchAllHabits = () => {
-  console.log('test');
-  return habitDAO.findAll();
+exports.fetchAllHabits = async () => {
+  data = await habitDAO.findAll();
+  if (data.length === 0)
+    throw new ElementNotFoundException('The habits list is empty')
+  return data;
 };
 
-exports.fetchHabit = (user, id) => {
-  console.log('test');
-  return habitDAO.findById(id);
+exports.fetchHabit = async (id) => {
+  data = await habitDAO.findById(id);
+  if (!data)
+    throw new ElementNotFoundException(`The habit ID ${id} doesn't exists`)
+  return data;
 };
 
-exports.deleteHabit = (user, id) => {
-  console.log('test');
+exports.deleteHabit = async (id) => {
+  data = await habitDAO.findById(id);
+  if (!data)
+    throw new ElementNotFoundException(`The habit ID ${id} doesn't exists`)
   return habitDAO.deleteById(id);
 };
 
-exports.createHabit = (user, data) => {
-  console.log('test');
-  return habitDAO.create(data);
-};
-
-exports.updateHabit = (user, data, id) => {
-  console.log('test');
+exports.updateHabit = async (data, id) => {
+  data = await habitDAO.findById(id);
+  if (!data)
+    throw new ElementNotFoundException(`The habit ID ${id} doesn't exists`)
   return habitDAO.updateHabit(data, id);
 };
 
+exports.createHabit = async (data) => {
+  return habitDAO.create(data);
+};
